requirements.txt
#psycopg2
Django==1.3.1
wsgiref==0.1.2
django-annoying